file cache creation of data for optimum performance. 

In my localhost, with a laravel project with ~300 routes, I reduce time about 10 to 56 ms