@bakerkretzmar 2024-02-20
Not totally convinced that this is necessary, but open to it if the performance gains are significant. I'm surprised that reading from the file and calling `collect()` is faster than doing it in memory.

Can you add tests for this and update your code comments to be in English? Also, is this tied to the route cache at all? How does it get cleared/updated when routes change?

Thanks!