@alhawaj-bh 2024-11-23
I conducted benchmarks for file cache route creation to evaluate performance improvements.
In our app with over 4000 routes, profiling revealed that nameKeyedRoutes() is the top function in terms of execution time.


### Results Before:

- CPU Time: 139ms
- Wall Time: 176ms
- Memory Usage: 5.77MB
- Peak Memory Usage: 6.40MB


### Results After:

- CPU Time: 52ms (62% improvement)
- Wall Time: 57ms (68% improvement)
- Memory Usage: 7.39MB
- Peak Memory Usage: 7.78MB



The changes significantly reduce CPU and wall time while slightly increasing memory usage.

![profiler-019358ad-bf3e-732b-8b0d-05613f07caee](https://github.com/user-attachments/assets/cecc35cf-72fd-43e1-adaa-9e77888dab2a)
![profiler-0193589e-fd30-705c-beca-0637e650059b (1)](https://github.com/user-attachments/assets/ab4d7d9f-1f71-4018-a99e-f022267370d2)