@msoler75 2025-11-19
I tested in production:

Standard ziggy constructor:
> php artisan app:profile-ziggy
10.95 ms

Constructor with json file cache:
> php artisan app:profile-ziggy
0.82 ms

My routes:
bootstrap/cache/routes-v7.php  size 904 kb
bootstrap/cache/ziggy2.json  size  54 kb  <-- this is the json file cache created


Code of console command profiler profileZiggy.php:

```
public function handle()
    {
        // we clear routes here
        new \Tightenco\Ziggy\Ziggy();
        \Tightenco\Ziggy\Ziggy::clearRoutes();

        // now we load constructor
        $start = microtime(true);
        $ziggy = new \Tightenco\Ziggy\Ziggy();
        $end = microtime(true);
        $time = ($end - $start) * 1000;
        $this->line(number_format($time, 2));
    }

```